drivers/sensor: lis2dux12: fix temperature conversion
Return to sensor_api i/f the temperature in Celsius instead
of the register raw value in LSB.

Fixes #75686

+++ drivers/sensor/st/lis2dux12/lis2dux12.h
@@ -55,7 +55,7 @@ struct lis2dux12_data {
 	uint8_t odr;
 
 #ifdef CONFIG_LIS2DUX12_ENABLE_TEMP
-	int sample_temp;
+	float sample_temp;
 #endif
 
 #ifdef CONFIG_LIS2DUX12_TRIGGER
